Add comprehensive integration tests for resource functions
Fixes:
- Added resource test functions to correct location (server/tests/)
- Added start_with_resources() helper to StdioServer
- Added create_temp_resource_dir() and cleanup_resource_dir() to common module
- Removed incorrectly placed test files in wrong directory

Tests added (8 comprehensive test cases):
- test_stdio_resource_write_and_read: Basic read/write operations
- test_stdio_resource_list: Directory listing functionality
- test_stdio_resource_delete: File deletion
- test_stdio_resource_nested_directories: Nested directory handling
- test_stdio_resource_error_handling: Error cases validation
- test_stdio_resource_file_uri_scheme: URI scheme support
- test_stdio_resource_json_operations: Complex JSON workflows

All tests properly create isolated temp directories for resource storage.
